Added phpstan dev dependency
Added composer dev dependency

Added phpstan script in composer

Changes in travis script:
Now phpstan is required in composer.json. In travis config PHPSTAN=0 by default and if it still 0 in particular build it will be removed from the dependencies and will not be installed and run.

Changes in appveyor:
Now phpstan exluded in appveyor because there used php below 7 version.