reverted issue #135 DST handling while date modifying
The approach was wrong. Because not all users expect the same behavior of modify in matter of handling DST changes. There is use cases for both approaches in both diapasones of units (less than day and equal or bigger). And by default php already have the way to handle this. You just should

* use UTC timezone if you don't want to have timestamp adjusments in case of DST changes
* use non-UTC timezone if you want to have such adjustments