Simplify ChronosInterval.
It now only provides a static method for creating it's instances
and __toString() to get the ISO 8601 interval string.

All methods which manipulate the properties of the extended
DateInterval have been dropped. The PHP manual states that they
should be considered readable only. An exception is made only for
DateInterval::$f as the constructor doesn't support specifying
microseconds through the interval string.