Make setTimezone deprecation conditional (#372)
Unfortunately the `assertEquals()` implementation for datetime values
calls `setTimezone()` on instances. Because, I don't think it is
reasonable to ask PHPUnit to have conditional logic for this, and
I don't think it is reasonable to ask users to adapt all of their test
code. These constraints led me to the solution of inspecting the
preceding frames to see if we're inside PHPUnit assertion. If we aren't
a deprecation is still emitted.